Move charge version metadata to NALD import job
