+++ src/modules/charging-import/jobs/index.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const IMPORT_CHARGING_DATA = 'import.charging-data';
-const IMPORT_CHARGE_VERSION_METADATA = 'import.charge-version-metadata';
 
 /**
  * Formats arguments to publish a PG boss event to import all companies
@@ -14,19 +13,5 @@ const importChargingData = () => ({
   }
 });
 
-/**
- * Formats arguments to publish a PG boss event to import charge version metadata
- * @return {Object}
- */
-const importChargeVersionMetadata = () => ({
-  name: IMPORT_CHARGE_VERSION_METADATA,
-  options: {
-    singletonKey: `${IMPORT_CHARGE_VERSION_METADATA}`
-  }
-});
-
 exports.IMPORT_CHARGING_DATA = IMPORT_CHARGING_DATA;
 exports.importChargingData = importChargingData;
-
-exports.IMPORT_CHARGE_VERSION_METADATA = IMPORT_CHARGE_VERSION_METADATA;
-exports.importChargeVersionMetadata = importChargeVersionMetadata;
