+++ src/modules/charging-import/plugin.js
@@ -2,7 +2,6 @@ const cron = require('node-cron');
 const jobs = require('./jobs');
 
 const chargingImport = require('./lib/import');
-const chargeVersionMetadataImport = require('./lib/import-charge-version-metadata');
 
 const { createRegister } = require('../../lib/plugin');
 const config = require('../../../config');
@@ -10,14 +9,9 @@ const config = require('../../../config');
 const registerSubscribers = async server => {
   // Register PG boss job handlers
   await server.messageQueue.subscribe(jobs.IMPORT_CHARGING_DATA, chargingImport.importChargingData);
-  await server.messageQueue.subscribe(jobs.IMPORT_CHARGE_VERSION_METADATA, chargeVersionMetadataImport.importChargeVersionMetadata);
 
   cron.schedule(config.import.charging.schedule,
     () => server.messageQueue.publish(jobs.importChargingData())
-  );
-
-  cron.schedule(config.import.chargeVersionMetadata.schedule,
-    () => server.messageQueue.publish(jobs.importChargeVersionMetadata())
   );
 };
