+++ src/modules/charging-import/routes.js
@@ -7,10 +7,5 @@ module.exports = [
     method: 'post',
     handler: controller.postImportChargingData,
     path: '/import/1.0/charging'
-  },
-  {
-    method: 'post',
-    handler: controller.postImportChargeVersionMetadata,
-    path: '/import/1.0/charge-version-metadata'
   }
 ];
